Move packaging commands into own folder
