+++ src/vpk/Velopack.Vpk/Commands/BaseCommand.cs
