+++ src/vpk/Velopack.Vpk/Commands/OutputCommand.cs
