+++ src/vpk/Velopack.Vpk/Commands/Packaging/LinuxPackCommand.cs
@@ -1,6 +1,6 @@
 ﻿using Velopack.Packaging.Unix;
 
-namespace Velopack.Vpk.Commands;
+namespace Velopack.Vpk.Commands.Packaging;
 
 public class LinuxPackCommand : PackCommand
 {
