+++ src/vpk/Velopack.Vpk/Commands/Packaging/OsxPackCommand.cs
@@ -1,4 +1,4 @@
-﻿namespace Velopack.Vpk.Commands;
+﻿namespace Velopack.Vpk.Commands.Packaging;
 
 public class OsxPackCommand : OsxBundleCommand
 {
