+++ src/vpk/Velopack.Vpk/Commands/PlatformCommand.cs
